requirements.txt
acme==0.31.0
aiohttp==3.7.4.post0
alabaster==0.7.12
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==21.2.0
awscli==1.16.113
Babel==2.9.1
blinker==1.4
boto==2.44.0
botocore==1.12.103
certbot==0.31.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
cloud-init==20.2
colorama==0.4.4
commonmark==0.9.1
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==2.6.1
discord==1.7.3
discord-ui==5.1.6
discord.py==1.7.3
distro-info==0.21
docker==6.0.0
docutils==0.17.1
entrypoints==0.3
future==0.16.0
httplib2==0.11.3
idna==3.3
imagesize==1.3.0
importlib-metadata==4.8.2
Jinja2==3.0.3
jmespath==0.9.4
josepy==1.1.0
jsonpatch==1.21
jsonpointer==1.10
jsonschema==2.6.0
keyring==17.1.1
keyrings.alt==3.1.1
livereload==2.6.3
Markdown==3.3.6
markdown-it-py==2.0.0
MarkupSafe==2.0.1
mdit-py-plugins==0.3.0
mdurl==0.1.0
mock==2.0.0
multidict==5.2.0
mysql-connector-python==8.0.30
myst-parser==0.16.0
oauthlib==2.1.0
packaging==21.3
parsedatetime==2.4
pbr==4.2.0
protobuf==3.20.1
pyasn1==0.4.2
pycrypto==2.6.1
pycryptodome==3.12.0
pycurl==7.43.0.2
Pygments==2.10.0
PyGObject==3.30.4
PyJWT==1.7.0
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyRFC3339==1.1
PySimpleSOAP==1.16.2
python-apt==1.8.4.3
python-dateutil==2.7.3
python-debian==0.1.35
python-debianbts==2.8.2
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
recommonmark==0.7.1
reportbug===7.5.3-deb10u1
requests==2.26.0
requests-toolbelt==0.8.0
roman==2.0.0
rsa==4.0
s3transfer==0.2.0
SecretStorage==2.3.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tornado==6.1
typing-extensions==4.0.1
unattended-upgrades==0.1
urllib3==1.26.7
websocket-client==1.4.0
yarl==1.7.2
zipp==3.6.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
